add phone number input dialog
